requirements.txt
django_rest_framework
django-cors-headers
django